requirements.txt
# Django Core
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Celery & Redis
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
PyJWT==2.8.0
email-validator==2.1.0
